refactor(t14): move HttpClientPort from core to adapter-binance
HttpClientPort is a transport detail needed only by Binance adapters — it has no
domain meaning in core. Moved to com.marmitt.binance.rest where it is consumed.
OkHttpClientAdapter in spring-application implements the port from adapter-binance
(spring-application already depends on adapter-binance, no new dependency added).

core/ now has zero HTTP transport ports.
