+++ adapter-binance/src/main/java/com/marmitt/binance/filters/SymbolFilterCache.java
@@ -2,7 +2,7 @@ package com.marmitt.binance.filters;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.marmitt.core.ports.outbound.http.HttpClientPort;
+import com.marmitt.binance.rest.HttpClientPort;
 import lombok.extern.slf4j.Slf4j;
 
 import java.io.IOException;
