+++ spring-application/src/main/java/com/marmitt/application/spring/adapter/binance/OkHttpClientAdapter.java
@@ -1,6 +1,6 @@
 package com.marmitt.application.spring.adapter.binance;
 
-import com.marmitt.core.ports.outbound.http.HttpClientPort;
+import com.marmitt.binance.rest.HttpClientPort;
 import okhttp3.MediaType;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
